[momo-MSM] 비정기 업데이트 detail: 2411111216
