--- momo/src/main/java/ks52team01/admin/dto/Admin.java
@@ -1,5 +1,0 @@
-package ks52team01.admin.dto;
-
-public class Admin {
-
-}
