+++ momo/src/main/java/ks52team01/admin/user/controller/AdminController.java
@@ -1,6 +1,7 @@
-package ks52team01.admin.controller;
+package ks52team01.admin.user.controller;
 
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 
 @Controller
