+++ momo/src/main/java/ks52team01/admin/user/dto/Admin.java
@@ -1,0 +1,5 @@
+package ks52team01.admin.user.dto;
+
+public class Admin {
+
+}
