+++ momo/src/main/java/ks52team01/admin/user/service/AdminService.java
@@ -1,25 +1,22 @@
-package ks52team01.admin.mapper;
+package ks52team01.admin.user.service;
 
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface AdminMapper {
+public interface AdminService {
 
 	public void adminPage();
-
-	public void adminLoginMove();
-
+	
+	public void adminLoginMove() ;
+	
 	public void adminLogoutAction();
-
+	
 	public void adminInfoModifyAction();
-
+	
 	public void adminJoinAction();
-
+	
 	public void adminInfoDetailAction();
-
+	
 	public void adminDeleteAction();
-
+	
 	public void adminFindIdAction();
-
+	
 	public void adminFindPwAction();
 }
