+++ momo/src/main/resources/templates/view/home/index.html
@@ -34,7 +34,9 @@
 				<div class="categories__slider owl-carousel">
 					<div class="col-lg-3">
 						<div id="categoriesExamList" class="categories__item categories-div">
+							<div style="background-color: white;">
 							<h5><a class="categories-site-map-a" href="/exam/examList">모의고사 목록</a></h5>
+							</div>
 						</div>
 					</div>
 					<div class="col-lg-3">
